feat(media): add auth guards

+++ src/routes/mediaItem/mediaItem.service.ts
@@ -18,10 +18,32 @@ export class MediaItemService {
     private readonly mediaItemRepository: MediaItemRepositoryInterface,
   ) {}
 
+  private async isMediaItemOwner(
+    id: string,
+    userId: string,
+    mediaItemBase?: MediaItemDto,
+  ) {
+    const mediaItem =
+      mediaItemBase ?? (await this.mediaItemRepository.getMediaItemById(id));
+    const mediaList = await this.mediaListRepository.getMedialListById(
+      mediaItem.mediaListId,
+    );
+
+    if (!mediaItem || !mediaList) {
+      throw new HttpException(
+        `Media item with id '${id}' doesn't exist.`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
+
+    return mediaList.userId === userId;
+  }
+
   async createMediaItem(
     mediaId: number,
     mediaType: MediaItemDto['mediaType'],
     mediaListId: string,
+    userId: string,
   ) {
     const mediaList =
       await this.mediaListRepository.getMedialListById(mediaListId);
@@ -33,6 +55,10 @@ export class MediaItemService {
       );
     }
 
+    if (mediaList.userId !== userId) {
+      throw new HttpException('Unauthorized.', HttpStatus.UNAUTHORIZED);
+    }
+
     return this.mediaItemRepository.createMediaItem(
       mediaId,
       mediaType,
@@ -40,22 +66,38 @@ export class MediaItemService {
     );
   }
 
-  async getMediaItemsByListId(mediaListId: string) {
+  async getMediaItemsByListId(mediaListId: string, userId: string) {
+    const mediaList =
+      await this.mediaListRepository.getMedialListById(mediaListId);
+
+    if (mediaList && mediaList.userId !== userId && !mediaList.isPublic) {
+      throw new HttpException(`Unauthorized.`, HttpStatus.UNAUTHORIZED);
+    }
+
     return this.mediaItemRepository.getMediaItemsByListId(mediaListId);
   }
 
-  async getMediaItemById(id: string) {
-    return this.mediaItemRepository.getMediaItemById(id);
-  }
+  async deleteMediaItem(id: string, userId: string) {
+    const isMediaItemOwner = await this.isMediaItemOwner(id, userId);
 
-  async deleteMediaItem(id: string) {
+    if (!isMediaItemOwner) {
+      throw new HttpException('Unauthorized.', HttpStatus.UNAUTHORIZED);
+    }
+
     return this.mediaItemRepository.deleteMediaItem(id);
   }
 
   async updateMediaItemTrackingData(
     id: string,
     trackingData: MediaItemDto['trackingData'],
+    userId: string,
   ) {
+    const isMediaItemOwner = await this.isMediaItemOwner(id, userId);
+
+    if (!isMediaItemOwner) {
+      throw new HttpException('Unauthorized.', HttpStatus.UNAUTHORIZED);
+    }
+
     return this.mediaItemRepository.updateMediaItemTrackingData(
       id,
       trackingData,
